trend-scanner/pyproject.toml
[build-system]
requires = ["setuptools>=68"]
build-backend = "setuptools.build_meta"

[project]
name = "trend-scanner"
version = "0.1.0"
description = "Reddit-only Trend Scanner MVP"
requires-python = ">=3.11"

dependencies = [
    "python-dotenv>=1.0.1",
    "pydantic>=2.7.0",
    "pydantic-settings>=2.3.0",
    "sqlalchemy>=2.0.30",
    "requests>=2.32.0",
    "rich>=13.7.1",
]

[tool.ruff]
line-length = 100
target-version = "py311"
